+++ src/main.ts
@@ -1,4 +1,6 @@
-//INFO: HTTPSApi Не верно возвращает данные при ошибке всего лишь 
+//INFO: HTTPSApi Не верно возвращает данные при ошибке всего лишь
+
+// import { Color } from "@lib";
 
 // import { Color } from '@lib';
 
@@ -28,7 +30,7 @@
 //       // }else{
 //       //   reject({ error: 'Нет данных', meta: meta });
 //       // }
-//     } catch (error) { 
+//     } catch (error) {
 //       console.log("HTTPSApiRequest error", error);
 //       const { msg, ...otherErr } = error as ResponseErrorHTTPSApi;
 //       reject({ error: msg, meta: { ...otherErr, msg } });
@@ -36,3 +38,25 @@
 //     }
 //   });
 // }
+
+// document.addEventListener("DOMContentLoaded", () => {
+//   Array.from({ length: 10 }, () => {
+//     const div = document.createElement("div");
+//     document.body.style.height = "100vh";
+//     const bg = Color.generatePleasantColor({
+//       //   saturation: 50, // Увеличиваем насыщенность
+//       //   lightness: 90, // Делаем ярче
+//       lightness: 95, // Делаем ярче
+//       saturation: 65, // Увеличиваем насыщенность
+//       hueShift: 0,
+//       //   randomize: true
+//     });
+//     console.log(bg);
+//     div.style.backgroundColor = bg;
+//     div.style.width = "100%";
+//     div.style.marginTop = "5px";
+//     div.style.height = "20px";
+//     document.body.appendChild(div);
+//     //   el.appendChild(div);
+//   });
+// });